Skeleton building
Not much today but I have a reason for that. I tried to get ESLint
working, but it was yelling at me for having CRLF endings on EVERY line
and I told him, that's what I want, but he kept fucking doing it so
after almost 2 hours I was back to no linter.